+++ src/configs/index.ts
@@ -1,5 +1,6 @@
 import { appConfig } from '@/configs/app.config';
 import { dbConfig } from '@/configs/db.config';
+import { mediamtxConfig } from '@/configs/mediamtx.config';
 import { mqttConfig } from '@/configs/mqtt.config';
 import { redisConfig } from '@/configs/redis.config';
 
@@ -8,4 +9,5 @@ export const config = {
     db: dbConfig,
     redis: redisConfig,
     mqtt: mqttConfig,
+    mediamtx: mediamtxConfig,
 } as const;